translate/requirements.txt
# BMAD Project Requirements

# Translation System
argostranslate>=1.9.0
pyyaml>=6.0
tomli>=2.0.1; python_version < "3.11"
tomli-w>=1.0.0
language-tool-python
langdetect
tqdm
